requirements.txt
# Medical RAG System Requirements

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1

# ML/NLP libraries
numpy==1.24.3
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
scispacy==0.5.3
spacy==3.7.2

# Vector search
faiss-cpu==1.7.4
# For GPU: faiss-gpu==1.7.4

# Search/IR
elasticsearch==8.11.0
rank-bm25==0.2.2

# LLM clients
openai==1.3.0

# Evaluation
rouge-score==0.1.2

# Utilities
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0

# Jupyter (for evaluation)
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.0
matplotlib==3.8.2
